refactor(temporal): more optimization of jiff extension structure
